Rename modules (#150)
* Rename modules

* Rename modules and reformat
